Rename akka fork join instrumentation package
