+++ instrumentation/akka/akka-actor-fork-join-2.5/javaagent/src/main/java/io/opentelemetry/javaagent/instrumentation/akkaforkjoin/VirtualFields.java
@@ -3,7 +3,7 @@
  * SPDX-License-Identifier: Apache-2.0
  */
 
-package io.opentelemetry.javaagent.instrumentation.akkaactor;
+package io.opentelemetry.javaagent.instrumentation.akkaforkjoin;
 
 import akka.dispatch.forkjoin.ForkJoinTask;
 import io.opentelemetry.instrumentation.api.util.VirtualField;
